Tag check if already deleted before reattempting delete

+++ src/main/java/com/company/web/forum/services/TagServiceImpl.java
@@ -1,6 +1,7 @@
 package com.company.web.forum.services;
 
 import com.company.web.forum.exceptions.AuthorizationException;
+import com.company.web.forum.exceptions.EntityDeletedException;
 import com.company.web.forum.exceptions.EntityDuplicateException;
 import com.company.web.forum.exceptions.EntityNotFoundException;
 import com.company.web.forum.models.*;
@@ -79,8 +80,11 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void delete(int id, User user) {
+        Tag tag = tagRepository.get(id);
         checkModifyPermissions(id, user);
-        //check if tag already deleted?
+        if (tag.getIsDeleted() == 1) {
+            throw new EntityDeletedException("Tag", "ID", String.valueOf(id));
+        }
         tagRepository.delete(id);
     }
